backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pandas==2.2.0
python-multipart==0.0.6
openai>=1.12.0
python-dotenv==1.0.1
aiofiles==23.2.1
scikit-learn>=1.4.0
joblib>=1.3.0

# Authentication
PyJWT>=2.8.0
httpx>=0.26.0
cryptography>=42.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0